long 타입을 Long 타입으로 변경

+++ src/main/java/apiMonitering/domain/ApiResponse.java
@@ -23,7 +23,7 @@ public class ApiResponse {
 
     private int statusCode;
 
-    private int responseTime;
+    private Long responseTime;
 
     private String body;
